refactored larger areas of code to components
