+++ src/lib/components/Slot/WinBanner.svelte
@@ -1,0 +1,8 @@
+<div class="bg-white relative h-[50px] w-[230px] center rounded-lg">
+	<img
+		src="/pikachu-happy.svg"
+		alt="pikachu-happy"
+		class="h-[250px] absolute -left-[10px] -top-[30px]"
+	/>
+	<p class="bg-white px-5 rounded-lg text-3xl text-right w-full font-bold">WINNER!</p>
+</div>
